refactor extract test suite info
